Add ATPneumaticsCheckout unit test. Merge all DayTime Checkout unit tests into one module; test_auxtel_daytime_checkouts.py. Remove SalObj import as it's no longer used.
